+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- Better handling of `/etc/asusd` not existing
+- Better handling of non-existant config files
 - Added option to set `disable_nvidia_powerd_on_battery`
 - Add short log entry to throttle_thermal_policy change detection
 - ROGCC: Don't notify user if changing to same mux mode